Learning about strings and bytes a bit
io.Reader/io.Writer seem like a really important part
of the entire golang ecosystem, so learning it seems
crucial.

Here I'm trying different ways of reading a multiline
string literal.

First, reading it byte by byte until the EOF.

Then trying two different approaches with bufio's
ReadByte and ReadString.

Finally rewriting the ReadString approach using a closure.
I like that version, but I'm curious is there a way to have
the call to the `readLine()` in the `for` loop definition,
so that when it returns a `false`, the `for` ends.